requirements.txt
Flask==3.0.3
Flask_SQLAlchemy==3.1.1
python-dotenv==1.0.1
reportlab==4.2.5
qrcode==7.4.2
Pillow==10.4.0
